Resolve issue with both the unordered list duplicating with each addition to it as well as the removal of lists leaving a null value in the local storage object.
Regarding the first issue, the list was not being clearned with a new addition to the list. Therefore, with each addition, the data set was being looped through and each item was getting appended fresh. This was cleared by using the removeChild() method.

Regarding the last issue, the lists data set residing in the user object was being treated as an array even through it was technically an object. Application of the delete key word was resulting in null values in the data set which would crash the app. The splice() method was instead applied for desire affect.