fix: implement .trim() method on list inputs
There was a bug that prevented editing of list titles when a trailing
space was inserted at the end. This was because the space would change
what was in local storage vs what was in the input field, thus breaking
the function used to identify data. The .trim() function was applied
to counteract trailing spaces.