fix: Correct variable scope issue with back button
When enabling back button functionality between todo page and list page,
a bug was discovered when the todo container would not be removed.
An error was shown that the map was null, because the variable that had
the mapped data was a local variable. Changing the scope of the variable
fixed the issue.